backend/app/apps/parties/urls.py
from rest_framework.routers import DefaultRouter

from app.apps.parties.views import PartyViewSet

router = DefaultRouter()
router.register("parties", PartyViewSet, basename="party")

urlpatterns = router.urls
