backend/pyproject.toml
[project]
name = "election-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<5.2",
    "djangorestframework>=3.15,<4",
    "psycopg[binary]>=3.2,<4",
    "django-filter>=24.3,<25",
    "drf-spectacular>=0.28,<1",
    "django-cors-headers>=4.6,<5",
    "gunicorn>=23,<24",
]

[dependency-groups]
dev = [
    "django-stubs>=5.1,<6",
    "djangorestframework-stubs>=3.15,<4",
    "pyright>=1.1.390",
    "ruff>=0.8",
    "pre-commit>=4",
]

[tool.ruff]
target-version = "py312"
line-length = 99
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["RUF012"]

[tool.django-stubs]
django_settings_module = "app.settings.dev"
